+++ app/Views/users/login.php
@@ -1,6 +1,6 @@
 <?php require_once $_SERVER['DOCUMENT_ROOT'].'/../app/Views/layouts/header.php'; ?>
 <?php 
-if (!isset($_SESSION['userLogin'])) {
+if (isset($_SESSION['userLogin'])) {
   header('location: /home/index');
 }
 ?>
@@ -11,7 +11,7 @@ if (!isset($_SESSION['userLogin'])) {
     <div class="row">
       <div class="col-lg-8 col-md-10 mx-auto">
         <div class="page-heading">
-          <h1>New Blog</h1>
+          <h1>Login</h1>
           <!-- <span class="subheading">Have questions? I have answers.</span> -->
         </div>
       </div>
@@ -27,25 +27,18 @@ if (!isset($_SESSION['userLogin'])) {
       <!-- Contact Form - Enter your email address on line 19 of the mail/contact_me.php file to make this form work. -->
       <!-- WARNING: Some web hosts do not allow emails to be sent through forms to common mail hosts like Gmail or Yahoo. It's recommended that you use a private domain email address! -->
       <!-- To use the contact form, your site must be on a live web host with PHP! The form will not work locally! -->
-      <form action="/home/addBlog" name="sentMessage" id="contactForm" novalidate method="POST">
+      <form action="/users/postLogin" name="sentMessage" id="contactForm" novalidate method="POST">
         <div class="control-group">
           <div class="form-group floating-label-form-group controls">
-            <label>Name</label>
-            <input type="text" name="name" value="<?php echo isset($name) ? $name : ''; ?>" class="form-control" placeholder="Name" id="name" required>
+            <label>Username</label>
+            <input type="text" name="username" value="<?php echo isset($username) ? $username : ''; ?>" class="form-control" placeholder="Username" id="name" required>
             <p class="help-block text-danger"></p>
           </div>
         </div>
         <div class="control-group">
           <div class="form-group floating-label-form-group controls">
-            <label>Preview text</label>
-            <textarea name="preview_text" rows="3" class="form-control" placeholder="Preview text" required></textarea>
-            <p class="help-block text-danger"></p>
-          </div>
-        </div>
-        <div class="control-group">
-          <div class="form-group floating-label-form-group controls">
-            <label>Detail text</label>
-            <textarea name="detail_text" rows="10" class="form-control" placeholder="Detail text" required></textarea>
+            <label>Password</label>
+            <input type="password" name="password" value="<?php echo isset($password) ? $password : ''; ?>" class="form-control" placeholder="Password" id="email" required>
             <p class="help-block text-danger"></p>
           </div>
         </div>
@@ -59,7 +52,8 @@ if (!isset($_SESSION['userLogin'])) {
         ?>
         </div>
         <div class="form-group">
-          <button type="submit" name="btnLogin" class="btn btn-primary" id="sendMessageButton">Post</button>
+          <button type="submit" name="btnLogin" class="btn btn-primary" id="sendMessageButton">Login</button>
+          <p><a href="/users/register">Register!</a></p>
         </div>
       </form>
     </div>
