--- app/Views/users/show.php
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <title>Show user</title>
-    </head>
-    <body>
-        <h1><?php echo $username . ' - ' . $id; ?></h1>
-    </body>
-</html>
