--- app/Views/users/test.php
@@ -1,1 +1,0 @@
-<h3>function test of users</h3>
